+++ components/layout/construct.rs
@@ -52,8 +52,8 @@ use style::context::SharedStyleContext;
 use style::logical_geometry::Direction;
 use style::properties::{self, ServoComputedValues};
 use style::selector_impl::{PseudoElement, RestyleDamage};
-use style::selector_matching::Stylist;
 use style::servo::restyle_damage::{BUBBLE_ISIZES, RECONSTRUCT_FLOW};
+use style::stylist::Stylist;
 use table::TableFlow;
 use table_caption::TableCaptionFlow;
 use table_cell::TableCellFlow;
